Python 3.
1. Get rid of circular reference to blob.
2. Ran 2to3 over codebase, commit changes.
