+++ lastpass/fetcher.py
@@ -1,12 +1,12 @@
 # coding: utf-8
-import httplib
+import http.client
 import pbkdf2
 import hashlib
 import requests
 #from lxml import etree
 from xml.etree import ElementTree as etree
-import blob 
-from exceptions import (
+
+from lastpass.exceptions import (
     NetworkError,
     InvalidResponseError,
     UnknownResponseSchemaError,
@@ -16,7 +16,19 @@ from exceptions import (
     LastPassIncorrectYubikeyPasswordError,
     LastPassUnknownError
 )
-from session import Session
+from lastpass.session import Session
+
+
+class Blob(object):
+    def __init__(self, bytes, key_iteration_count):
+        self.bytes = bytes
+        self.key_iteration_count = key_iteration_count
+
+    def encryption_key(self, username, password):
+        return fetcher.Fetcher.make_key(username, password, self.key_iteration_count)
+
+    def __eq__(self, other):
+        return self.bytes == other.bytes and self.key_iteration_count == other.key_iteration_count
 
 
 class Fetcher(object):
@@ -30,16 +42,16 @@ class Fetcher(object):
         response = web_client.get('https://lastpass.com/getaccts.php?mobile=1&b64=1&hash=0.0',
                                   cookies={'PHPSESSID': session.id})
 
-        if response.status_code != httplib.OK:
+        if response.status_code != http.client.OK:
             raise NetworkError()
 
-        return blob.Blob(cls.decode_blob(response.content), session.key_iteration_count)
+        return Blob(cls.decode_blob(response.content), session.key_iteration_count)
 
     @classmethod
     def request_iteration_count(cls, username, web_client=requests):
         response = web_client.post('https://lastpass.com/iterations.php',
                                    data={'email': username})
-        if response.status_code != httplib.OK:
+        if response.status_code != http.client.OK:
             raise NetworkError()
 
         try:
@@ -68,7 +80,7 @@ class Fetcher(object):
         response = web_client.post('https://lastpass.com/login.php',
                                    data=body)
 
-        if response.status_code != httplib.OK:
+        if response.status_code != http.client.OK:
             raise NetworkError()
 
         try:
@@ -88,7 +100,7 @@ class Fetcher(object):
     def create_session(cls, parsed_response, key_iteration_count):
         if parsed_response.tag == 'ok':
             session_id = parsed_response.attrib.get('sessionid')
-            if isinstance(session_id, basestring):
+            if isinstance(session_id, str):
                 return Session(session_id, key_iteration_count)
 
     @classmethod
@@ -119,7 +131,7 @@ class Fetcher(object):
     @classmethod
     def make_key(cls, username, password, key_iteration_count):
         if key_iteration_count == 1:
-            return hashlib.sha256(username + password).digest()
+            return hashlib.sha256((username.encode() + password.encode())).digest()
         else:
             return pbkdf2.pbkdf2_bin(password, username, key_iteration_count, 32, hashlib.sha256)
